src/dlstp/tp1/c_correlations.py
import pandas as pd


def compute_all_stock_cross_correlations(stock_prices: pd.DataFrame) -> pd.DataFrame:
  """
  Find cross correlations between stocks based on their price time series.

  Given a dataframe with columns date, symbol and close, find the overall correlation
  between all pairs of stocks based on the dynamics of their prices.

  You should return a dataframe whose index and columns are the stock symbols, and
  whose values are the correlations between each pair of stock.

  Tips:
      - understand what 'pivoting' a table means (see e.g. pd.pivot)
      - this can be done in a single line of code

  Notes
      - expect this to be a bit slow, as it requires calculating all stock pairs
      correlations

  """
  assert all(c in stock_prices.columns for c in ("close", "date", "symbol"))
  raise NotImplementedError()


def compute_monthly_stock_cross_correlations(
  stock_prices: pd.DataFrame,
) -> pd.DataFrame:
  """
  Great, but our data is _sequential_ in nature.

  How about looking at the evolving dynamics of these stock prices?

  How is their cross correlation evolving over time?

  Use the previous function, but applying it to each month of data instead of the
  entire dataset.

  This function should return a dataframe with 4 columns:
      - date (first day of the month for the time period)
      - symbol_a
      - symbol_b
      - corrcoef (correlation coefficient between -1 and 1)

  Tips:
      - look at pd.Grouper
      - the offset alias for 'start of the month' is 'MS' in pandas [1]
      - the opposit of 'pivotting' is 'melting' (pd.melt)

  [1]: https://pandas.pydata.org/pandas-docs/stable/user_guide/timeseries.html#offset-aliases

  Notes:
      - the number of price observations per month is not the same due to missing
        data, holidays. this is not a problem for estimating the correlation
      - sometimes a pair of symbols is not observed on a given month, this should
        result in NaN values in the resulting correlation matrix. not a problem

  """
  raise NotImplementedError()
